+++ CHANGELOG.markdown
@@ -4,6 +4,9 @@
 * Made `holesOf` and `partsOf` lazier to deal with infinite structures.
 * Resolved issue #75. We now generate nicer core for most `Setter` and `Fold` operations, and some others.
 * Made lenses for field access like `_1`, `_2`, etc. lazier.
+* Added `Control.Lens.Loupe`, which provides a limited form of `Lens` that can be read from and written to and which can compose
+  with other lenses, but can also be returned in a list or as a monadic result, but cannot be used directly for most combinators
+  without cloning it first. It is easier to compose than a `ReifiedLens`, but slightly slower.
 
 3.1
 ---